% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSQL.R
\name{getSQL}
\alias{getSQL}
\title{getSQL}
\usage{
getSQL(filepath)
}
\arguments{
\item{filepath}{the file path to a .sql file you would like to use}
}
\value{
The text from the .sql file
}
\description{
this makes getting sql easier into R, use this instead of copying the sql file
Just make sure to SET NOCOUNT ON if you are using temp tables
}
\examples{
getSQL("blep.sql")
}
