% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_to_sql.R
\name{string_to_sql}
\alias{string_to_sql}
\title{Convert array to sql code}
\usage{
string_to_sql(string)
}
\arguments{
\item{string}{An array or list of items you wish to delimit.}
}
\value{
A list delimeted by , and each block surrounded by ' '.. This is auto copied to your clip board.
}
\description{
This allows you to copy and array to your clipboard as a comma delimited and surrounded by string stuff.
This makes it easy to just copy lists of things for quick SQL queries.
}
\examples{
string_to_sql(mtcars$mpg)
}
